Add support for saving and loading sessions
Adds methods for saving and loading a session, to avoid having to
authenticate each time the API is used in a new execution, kernel
instance, etc.  Both the session cookies and CSRF token are stored to a
file via Python's bulit-in pickling.

Both methods take a file name param, to define where the data should be
written or read.